Move RestClient classes into separate files :truck:
