+++ Src/Notion.Client/RestClient/ClientOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace Notion.Client
+{
+    public class ClientOptions
+    {
+        public string BaseUrl { get; set; }
+        public string NotionVersion { get; set; }
+        public string AuthToken { get; set; }
+    }
+}
