+++ Src/Notion.Client/RestClient/IRestClient.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public interface IRestClient
+    {
+        Task<T> GetAsync<T>(
+            string uri,
+            IDictionary<string, string> queryParams = null,
+            IDictionary<string, string> headers = null,
+            JsonSerializerSettings serializerSettings = null,
+            CancellationToken cancellationToken = default);
+
+        Task<T> PostAsync<T>(
+            string uri,
+            object body,
+            IDictionary<string, string> queryParams = null,
+            IDictionary<string, string> headers = null,
+            JsonSerializerSettings serializerSettings = null,
+            CancellationToken cancellationToken = default);
+
+        Task<T> PatchAsync<T>(
+            string uri,
+            object body,
+            IDictionary<string, string> queryParams = null,
+            IDictionary<string, string> headers = null,
+            JsonSerializerSettings serializerSettings = null,
+            CancellationToken cancellationToken = default);
+    }
+}
